refactor: packages refactoring
