+++ src/main/kotlin/org/danilopianini/plagiarismdetector/analyzer/technique/tokenization/java/JavaTokenTypesSupplier.kt
@@ -1,10 +1,11 @@
-package org.danilopianini.plagiarismdetector.analyzer.steps.tokenization
+package org.danilopianini.plagiarismdetector.analyzer.technique.tokenization.java
 
 import com.charleskorn.kaml.Yaml
 import kotlinx.serialization.builtins.ListSerializer
 import org.danilopianini.plagiarismdetector.analyzer.representation.token.LanguageTokenTypes
 import org.danilopianini.plagiarismdetector.analyzer.representation.token.TokenTypeImpl
 import org.danilopianini.plagiarismdetector.analyzer.representation.token.LanguageTokenTypesImpl
+import org.danilopianini.plagiarismdetector.analyzer.technique.tokenization.LanguageTokenTypesSupplier
 
 /**
  * A supplier of [LanguageTokenTypesImpl].
