+++ src/main/kotlin/org/danilopianini/plagiarismdetector/analyzer/technique/tokenization/java/JavaTokenizer.kt
@@ -1,12 +1,11 @@
-package org.danilopianini.plagiarismdetector.analyzer.steps.tokenization.java
+package org.danilopianini.plagiarismdetector.analyzer.technique.tokenization.java
 
 import com.github.javaparser.ast.CompilationUnit
 import com.github.javaparser.ast.Node
 import com.github.javaparser.ast.visitor.TreeVisitor
 import org.danilopianini.plagiarismdetector.analyzer.representation.token.Token
 import org.danilopianini.plagiarismdetector.analyzer.representation.token.TokenImpl
-import org.danilopianini.plagiarismdetector.analyzer.steps.StepHandler
-import org.danilopianini.plagiarismdetector.analyzer.steps.tokenization.JavaTokenTypesSupplier
+import org.danilopianini.plagiarismdetector.analyzer.StepHandler
 
 /**
  * A Java source file tokenizer.
